requirements-backend.txt
fastapi==0.135.1
uvicorn==0.30.1
pydantic==2.12.5
pydantic-settings==2.3.0
python-multipart==0.0.6
requests==2.32.5
prometheus_client==0.24.1
scikit-learn==1.8.0
google-genai==0.3.0
redis==7.3.0
python-json-logger==2.0.7
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-requests==0.45b0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
cryptography==41.0.7
python-dotenv==1.0.0
